худший коммит эвер, но ддл на "что нибудь показать" завтра. потом я буду все это разгребать и жалеть о куче кривых решений :D:D

+++ domains/users/model.py
@@ -1,12 +1,12 @@
 from database import Base
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import String, SmallInteger
+from sqlalchemy import String, Integer
 
 class User(Base):
     __tablename__ = "users"
-    id: Mapped[int] = mapped_column(SmallInteger(), primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer(), primary_key=True, autoincrement=True)
     login: Mapped[str] = mapped_column(String(32), unique=True)
     password: Mapped[str] = mapped_column(String(32))
     name: Mapped[str] = mapped_column(String(32))
-    section_id: Mapped[int] = mapped_column(SmallInteger())
+    section_id: Mapped[int] = mapped_column(Integer())
     group_name: Mapped[str] = mapped_column(String(32))
